gdit_api_extract/requirements.txt
bcrypt
botocore
certifi
cffi
charset-normalizer
cryptography
glog
idna
jmespath
lxml
paramiko<4.0.0
psutil
pycparser
PyNaCl
python-dateutil
python-gflags
python-gnupg
pytz
requests
six
urllib3
